chore: move all queries about goal into @/hooks/queries

+++ src/hooks/queries/useGoals.ts
@@ -3,6 +3,7 @@ import { useMutation, useQuery } from "@tanstack/react-query";
 import { supabase } from "@/integrations/supabase/client";
 import { useAuth } from "@/hooks/useAuth";
 import { useToast } from "@/hooks/use-toast";
+import { GoalFormData } from "@/form-dto/goals";
 
 export const useGoals = () => {
   const { user } = useAuth();
@@ -25,12 +26,14 @@ export const useGoals = () => {
 
 export const useDeleteGoal = () => {
   const { toast } = useToast();
+  const { user } = useAuth();
 
   return useMutation({
     mutationFn: async (id: number) => {
       const { error } = await supabase
         .from("goals")
         .delete()
+        .eq("user_id", user?.id)
         .eq("id", id);
 
       if (error) throw error;
@@ -39,6 +42,67 @@ export const useDeleteGoal = () => {
       toast({
         title: "Berhasil",
         description: "Goal berhasil dihapus",
+      });
+    },
+    onError: (error: any) => {
+      toast({
+        title: "Error",
+        description: error.message,
+        variant: "destructive",
+      });
+    },
+  });
+};
+
+export const useUpdateGoal = () => {
+  const { toast } = useToast();
+  const { user } = useAuth();
+
+  return useMutation({
+    mutationFn: async ({ id, ...goal }: GoalFormData & { id: number }) => {
+      const { error } = await supabase
+        .from("goals")
+        .update(goal)
+        .eq("user_id", user?.id)
+        .eq("id", id);
+
+      if (error) throw error;
+    },
+    onSuccess: () => {
+      toast({
+        title: "Berhasil",
+        description: "Goal berhasil diperbarui",
+      });
+    },
+    onError: (error: any) => {
+      toast({
+        title: "Error",
+        description: error.message,
+        variant: "destructive",
+      });
+    },
+  });
+};
+
+export const useCreateGoal = () => {
+  const { toast } = useToast();
+  const { user } = useAuth();
+
+  return useMutation({
+    mutationFn: async (goal: GoalFormData) => {
+      const { error } = await supabase
+        .from("goals")
+        .insert({
+          ...goal,
+          user_id: user?.id,
+        });
+
+      if (error) throw error;
+    },
+    onSuccess: () => {
+      toast({
+        title: "Berhasil",
+        description: "Goal berhasil ditambahkan",
       });
     },
     onError: (error: any) => {
